# internal/apiserver/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  core.ErrResponse:
    properties:
      code:
        description: Code 定义了业务错误代码
        type: integer
      message:
        description: |-
          Message 包含此消息的详细信息
          此消息适合暴露给外部
        type: string
      reference:
        description: Reference 返回参考文档，可能有助于解决此错误
        type: string
    type: object
  core.Response:
    properties:
      code:
        description: Code 业务状态码，0 表示成功
        type: integer
      data:
        description: Data 响应数据
      message:
        description: Message 响应消息
        type: string
      reference:
        description: Reference 返回参考文档（错误时使用）
        type: string
    type: object
...
  statistics.AssessmentEntryStatisticsList:
    properties:
      items:
        items:
          $ref: '#/definitions/statistics.AssessmentEntryStatistics'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  statistics.ClinicianStatisticsList:
    properties:
      items:
        items:
          $ref: '#/definitions/statistics.ClinicianStatistics'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
info:
  contact: {}
  description: Questionnaire Scale API server (REST & gRPC)
  title: QS API Server
  version: "1.0"
paths:
...
  /api/v1/statistics/clinicians:
    get:
      consumes:
      - application/json
      description: 按机构和时间窗口查询从业者统计列表；仅 qs:admin 可访问
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 时间窗口预设：today/7d/30d
        in: query
        name: preset
        type: string
      - description: 自定义开始日期，格式 YYYY-MM-DD
        in: query
        name: from
        type: string
      - description: 自定义结束日期，格式 YYYY-MM-DD
        in: query
        name: to
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/core.Response'
            - properties:
                data:
                  $ref: '#/definitions/statistics.ClinicianStatisticsList'
              type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/core.ErrResponse'
      summary: 查询从业者统计列表
      tags:
      - Statistics
  /api/v1/statistics/clinicians/{id}:
    get:
      consumes:
      - application/json
      description: 按从业者 ID 查询统计概览、窗口指标和漏斗指标；仅 qs:admin 可访问
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 从业者ID
        in: path
        name: id
        required: true
        type: integer
      - description: 时间窗口预设：today/7d/30d
        in: query
        name: preset
        type: string
      - description: 自定义开始日期，格式 YYYY-MM-DD
        in: query
        name: from
        type: string
      - description: 自定义结束日期，格式 YYYY-MM-DD
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/core.Response'
            - properties:
                data:
                  $ref: '#/definitions/statistics.ClinicianStatistics'
              type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/core.ErrResponse'
      summary: 查询单个从业者统计
      tags:
      - Statistics
  /api/v1/statistics/clinicians/me/entries:
    get:
      consumes:
      - application/json
      description: 查询当前后台操作者绑定从业者的测评入口统计列表
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 时间窗口预设：today/7d/30d
        in: query
        name: preset
        type: string
      - description: 自定义开始日期，格式 YYYY-MM-DD
        in: query
        name: from
        type: string
      - description: 自定义结束日期，格式 YYYY-MM-DD
        in: query
        name: to
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/core.Response'
            - properties:
                data:
                  $ref: '#/definitions/statistics.AssessmentEntryStatisticsList'
              type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/core.ErrResponse'
      summary: 查询当前从业者入口统计列表
      tags:
      - Statistics
  /api/v1/statistics/clinicians/me/overview:
    get:
      consumes:
      - application/json
      description: 查询当前后台操作者绑定从业者的统计概览
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 时间窗口预设：today/7d/30d
        in: query
        name: preset
        type: string
      - description: 自定义开始日期，格式 YYYY-MM-DD
        in: query
        name: from
        type: string
      - description: 自定义结束日期，格式 YYYY-MM-DD
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/core.Response'
            - properties:
                data:
                  $ref: '#/definitions/statistics.ClinicianStatistics'
              type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/core.ErrResponse'
      summary: 查询当前从业者统计概览
      tags:
      - Statistics
  /api/v1/statistics/clinicians/me/testees-summary:
    get:
      consumes:
      - application/json
      description: 查询当前后台操作者绑定从业者可访问受试者摘要统计
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 时间窗口预设：today/7d/30d
        in: query
        name: preset
        type: string
      - description: 自定义开始日期，格式 YYYY-MM-DD
        in: query
        name: from
        type: string
      - description: 自定义结束日期，格式 YYYY-MM-DD
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/core.Response'
            - properties:
                data:
                  $ref: '#/definitions/statistics.ClinicianTesteeSummaryStatistics'
              type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/core.ErrResponse'
      summary: 查询当前从业者受试者摘要
      tags:
      - Statistics
  /api/v1/statistics/contents/batch:
    post:
      consumes:
      - application/json
      description: 按内容 type + code 批量查询测评形成数、完成数和完成率；questionnaire 需要 manage_questionnaires，scale 需要 manage_assessment_models，混合请求需要同时具备两项权限
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 内容标识列表
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.contentBatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/core.Response'
            - properties:
                data:
                  $ref: '#/definitions/statistics.ContentBatchStatisticsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/core.ErrResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/core.ErrResponse'
      summary: 批量查询统一内容统计
      tags:
      - Statistics
  /api/v1/statistics/entries:
    get:
      consumes:
      - application/json
      description: 查询机构内测评入口统计列表，支持 clinician_id、status 和时间窗口过滤；仅 qs:admin 可访问
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 从业者ID
        in: query
        name: clinician_id
        type: integer
      - description: 入口状态：active/inactive
        in: query
        name: status
        type: string
      - description: 时间窗口预设：today/7d/30d
        in: query
        name: preset
        type: string
      - description: 自定义开始日期，格式 YYYY-MM-DD
        in: query
        name: from
        type: string
      - description: 自定义结束日期，格式 YYYY-MM-DD
        in: query
        name: to
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/core.Response'
            - properties:
                data:
                  $ref: '#/definitions/statistics.AssessmentEntryStatisticsList'
              type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/core.ErrResponse'
      summary: 查询测评入口统计列表
      tags:
      - Statistics
  /api/v1/statistics/entries/{id}:
    get:
      consumes:
      - application/json
      description: 按入口 ID 查询入口快照、窗口计数和最近事件时间；仅 qs:admin 可访问
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 入口ID
        in: path
        name: id
        required: true
        type: integer
      - description: 时间窗口预设：today/7d/30d
        in: query
        name: preset
        type: string
      - description: 自定义开始日期，格式 YYYY-MM-DD
        in: query
        name: from
        type: string
      - description: 自定义结束日期，格式 YYYY-MM-DD
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/core.Response'
            - properties:
                data:
                  $ref: '#/definitions/statistics.AssessmentEntryStatistics'
              type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/core.ErrResponse'
      summary: 查询单个测评入口统计
      tags:
      - Statistics
  /api/v1/statistics/overview:
    get:
      consumes:
      - application/json
      description: 查询机构总览、接入漏斗、测评服务、维度分析和计划任务统计；仅 qs:admin 可访问
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 时间窗口预设：today/7d/30d
        in: query
        name: preset
        type: string
      - description: 自定义开始日期，格式 YYYY-MM-DD
        in: query
        name: from
        type: string
      - description: 自定义结束日期，格式 YYYY-MM-DD
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/core.Response'
            - properties:
                data:
                  $ref: '#/definitions/statistics.StatisticsOverview'
              type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/core.ErrResponse'
      summary: 查询机构统计总览
      tags:
      - Statistics
...
  /internal/v1/cache/governance/hotset:
    get:
      description: 按 kind 和 limit 查询当前缓存热集，仅内部管理员可访问。
      parameters:
      - description: 缓存类型
        in: query
        name: kind
        type: string
      - description: 最大条数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Response'
      summary: 查询缓存热集
      tags:
      - Statistics-Sync
...
schemes:
- http
- https
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"