# internal/apiserver/transport/rest/handler/statistics.go
package handler

import (
	"strconv"
	"strings"
	"time"

	"github.com/FangcunMount/component-base/pkg/errors"
	"github.com/FangcunMount/component-base/pkg/logger"
	authzApp "github.com/FangcunMount/qs-server/internal/apiserver/application/authz"
	statisticsApp "github.com/FangcunMount/qs-server/internal/apiserver/application/statistics"
	cachemodel "github.com/FangcunMount/qs-server/internal/apiserver/cache/governance/model"
	cachetarget "github.com/FangcunMount/qs-server/internal/apiserver/cache/governance/target"
	domainStatistics "github.com/FangcunMount/qs-server/internal/apiserver/domain/statistics"
	"github.com/FangcunMount/qs-server/internal/apiserver/transport/rest/response"
	"github.com/FangcunMount/qs-server/internal/pkg/code"
	"github.com/gin-gonic/gin"
)

// StatisticsHandler 统计处理器
type StatisticsHandler struct {
	BaseHandler
	readService                  statisticsApp.ReadService
	periodicStatsService         statisticsApp.PeriodicStatsService
	syncService                  statisticsApp.StatisticsSyncService
	testeeAccessService          statisticsApp.TesteeAccessValidator
	warmupCoordinator            statisticsApp.WarmupCoordinator
	cacheGovernanceStatusService statisticsApp.GovernanceStatusReader
	governanceFacade             statisticsApp.GovernanceFacade
}

// NewStatisticsHandler 创建统计处理器
func NewStatisticsHandler(
	readService statisticsApp.ReadService,
	periodicStatsService statisticsApp.PeriodicStatsService,
	syncService statisticsApp.StatisticsSyncService,
) *StatisticsHandler {
	return &StatisticsHandler{
		readService:          readService,
		periodicStatsService: periodicStatsService,
		syncService:          syncService,
	}
}

// SetTesteeAccessService 设置 testee 访问控制服务。
func (h *StatisticsHandler) SetTesteeAccessService(testeeAccessService statisticsApp.TesteeAccessValidator) {
	h.testeeAccessService = testeeAccessService
}

func (h *StatisticsHandler) SetWarmupCoordinator(coordinator statisticsApp.WarmupCoordinator) {
	h.warmupCoordinator = coordinator
	h.governanceFacade = nil
}

func (h *StatisticsHandler) SetCacheGovernanceStatusService(service statisticsApp.GovernanceStatusReader) {
	h.cacheGovernanceStatusService = service
	h.governanceFacade = nil
}

func (h *StatisticsHandler) SetGovernanceFacade(facade statisticsApp.GovernanceFacade) {
	h.governanceFacade = facade
}

func (h *StatisticsHandler) bindJSON(c *gin.Context, req interface{}) bool {
	if err := c.ShouldBindJSON(req); err != nil {
		h.Error(c, errors.WithCode(code.ErrInvalidArgument, "invalid request body: %v", err))
		return false
	}
	return true
}

func (h *StatisticsHandler) parsePage(c *gin.Context) (int, int, error) {
	page := 1
	pageSize := 20
	if raw := c.Query("page"); raw != "" {
		value, err := strconv.Atoi(raw)
		if err != nil {
			return 0, 0, errors.WithCode(code.ErrInvalidArgument, "invalid page: %s", raw)
		}
		page = value
	}
	if raw := c.Query("page_size"); raw != "" {
		value, err := strconv.Atoi(raw)
		if err != nil {
			return 0, 0, errors.WithCode(code.ErrInvalidArgument, "invalid page_size: %s", raw)
		}
		pageSize = value
	}
	return page, pageSize, nil
}

func buildStatisticsQueryFilter(c *gin.Context) statisticsApp.QueryFilter {
	return statisticsApp.QueryFilter{
		Preset: c.Query("preset"),
		From:   c.Query("from"),
		To:     c.Query("to"),
	}
}

func parseStatisticsSyncDateRange(c *gin.Context) (statisticsApp.SyncDailyOptions, error) {
	startRaw := strings.TrimSpace(c.Query("start_date"))
	endRaw := strings.TrimSpace(c.Query("end_date"))
	if startRaw == "" && endRaw == "" {
		return statisticsApp.SyncDailyOptions{}, nil
	}
	if startRaw == "" || endRaw == "" {
		return statisticsApp.SyncDailyOptions{}, errors.WithCode(code.ErrInvalidArgument, "start_date 和 end_date 必须同时提供")
	}

	start, err := time.ParseInLocation("2006-01-02", startRaw, time.Local)
	if err != nil {
		return statisticsApp.SyncDailyOptions{}, errors.WithCode(code.ErrInvalidArgument, "start_date 格式无效，必须为 YYYY-MM-DD")
	}
	endInclusive, err := time.ParseInLocation("2006-01-02", endRaw, time.Local)
	if err != nil {
		return statisticsApp.SyncDailyOptions{}, errors.WithCode(code.ErrInvalidArgument, "end_date 格式无效，必须为 YYYY-MM-DD")
	}
	endExclusive := endInclusive.AddDate(0, 0, 1)
	if !start.Before(endExclusive) {
		return statisticsApp.SyncDailyOptions{}, errors.WithCode(code.ErrInvalidArgument, "start_date 不能晚于 end_date")
	}

	return statisticsApp.SyncDailyOptions{
		StartDate: &start,
		EndDate:   &endExclusive,
	}, nil
}

type contentBatchRequest struct {
	Items []domainStatistics.ContentReference `json:"items"`
}

// GetOverview 查询机构统计总览。
// @Summary 查询机构统计总览
// @Description 查询机构总览、接入漏斗、测评服务、维度分析和计划任务统计；仅 qs:admin 可访问
// @Tags Statistics
// @Accept json
// @Produce json
// @Param Authorization header string true "Bearer 用户令牌"
// @Param preset query string false "时间窗口预设：today/7d/30d"
// @Param from query string false "自定义开始日期，格式 YYYY-MM-DD"
// @Param to query string false "自定义结束日期，格式 YYYY-MM-DD"
// @Success 200 {object} core.Response{data=statistics.StatisticsOverview}
// @Failure 429 {object} core.ErrResponse
// @Router /api/v1/statistics/overview [get]
func (h *StatisticsHandler) GetOverview(c *gin.Context) {
	ctx := c.Request.Context()
	orgID, err := h.RequireProtectedOrgID(c)
	if err != nil {
		h.Error(c, err)
		return
	}
	stats, err := h.readService.GetOverview(ctx, orgID, buildStatisticsQueryFilter(c))
	if err != nil {
		h.Error(c, err)
		return
	}
	h.Success(c, stats)
}

// ListClinicianStatistics 查询从业者统计列表。
// @Summary 查询从业者统计列表
// @Description 按机构和时间窗口查询从业者统计列表；仅 qs:admin 可访问
// @Tags Statistics
// @Accept json
// @Produce json
// @Param Authorization header string true "Bearer 用户令牌"
// @Param preset query string false "时间窗口预设：today/7d/30d"
// @Param from query string false "自定义开始日期，格式 YYYY-MM-DD"
// @Param to query string false "自定义结束日期，格式 YYYY-MM-DD"
// @Param page query int false "页码"
// @Param page_size query int false "每页数量"
// @Success 200 {object} core.Response{data=statistics.ClinicianStatisticsList}
// @Failure 429 {object} core.ErrResponse
// @Router /api/v1/statistics/clinicians [get]
func (h *StatisticsHandler) ListClinicianStatistics(c *gin.Context) {
	ctx := c.Request.Context()
	orgID, err := h.RequireProtectedOrgID(c)
	if err != nil {
		h.Error(c, err)
		return
	}
	page, pageSize, err := h.parsePage(c)
	if err != nil {
		h.Error(c, err)
		return
	}
	stats, err := h.readService.ListClinicianStatistics(ctx, orgID, buildStatisticsQueryFilter(c), page, pageSize)
	if err != nil {
		h.Error(c, err)
		return
	}
	h.Success(c, stats)
}

// GetClinicianStatistics 查询单个从业者统计。
// @Summary 查询单个从业者统计
// @Description 按从业者 ID 查询统计概览、窗口指标和漏斗指标；仅 qs:admin 可访问
// @Tags Statistics
// @Accept json
// @Produce json
// @Param Authorization header string true "Bearer 用户令牌"
// @Param id path uint64 true "从业者ID"
// @Param preset query string false "时间窗口预设：today/7d/30d"
// @Param from query string false "自定义开始日期，格式 YYYY-MM-DD"
// @Param to query string false "自定义结束日期，格式 YYYY-MM-DD"
// @Success 200 {object} core.Response{data=statistics.ClinicianStatistics}
// @Failure 429 {object} core.ErrResponse
// @Router /api/v1/statistics/clinicians/{id} [get]
func (h *StatisticsHandler) GetClinicianStatistics(c *gin.Context) {
	ctx := c.Request.Context()
	orgID, err := h.RequireProtectedOrgID(c)
	if err != nil {
		h.Error(c, err)
		return
	}
	clinicianID, err := strconv.ParseUint(c.Param("id"), 10, 64)
	if err != nil {
		h.Error(c, errors.WithCode(code.ErrInvalidArgument, "invalid clinician id: %s", c.Param("id")))
		return
	}
	stats, err := h.readService.GetClinicianStatistics(ctx, orgID, clinicianID, buildStatisticsQueryFilter(c))
	if err != nil {
		h.Error(c, err)
		return
	}
	h.Success(c, stats)
}

// ListAssessmentEntryStatistics 查询测评入口统计列表。
// @Summary 查询测评入口统计列表
// @Description 查询机构内测评入口统计列表，支持 clinician_id、status 和时间窗口过滤；仅 qs:admin 可访问
// @Tags Statistics
// @Accept json
// @Produce json
// @Param Authorization header string true "Bearer 用户令牌"
// @Param clinician_id query uint64 false "从业者ID"
// @Param status query string false "入口状态：active/inactive"
// @Param preset query string false "时间窗口预设：today/7d/30d"
// @Param from query string false "自定义开始日期，格式 YYYY-MM-DD"
// @Param to query string false "自定义结束日期，格式 YYYY-MM-DD"
// @Param page query int false "页码"
// @Param page_size query int false "每页数量"
// @Success 200 {object} core.Response{data=statistics.AssessmentEntryStatisticsList}
// @Failure 429 {object} core.ErrResponse
// @Router /api/v1/statistics/entries [get]
func (h *StatisticsHandler) ListAssessmentEntryStatistics(c *gin.Context) {
	ctx := c.Request.Context()
	orgID, err := h.RequireProtectedOrgID(c)
	if err != nil {
		h.Error(c, err)
		return
	}
	page, pageSize, err := h.parsePage(c)
	if err != nil {
		h.Error(c, err)
		return
	}
	var clinicianID *uint64
	if raw := c.Query("clinician_id"); raw != "" {
		value, err := strconv.ParseUint(raw, 10, 64)
		if err != nil {
			h.Error(c, errors.WithCode(code.ErrInvalidArgument, "invalid clinician_id: %s", raw))
			return
		}
		clinicianID = &value
	}
	var activeOnly *bool
	if raw := c.Query("status"); raw != "" {
		value := raw == "active"
		activeOnly = &value
	}
	stats, err := h.readService.ListAssessmentEntryStatistics(ctx, orgID, clinicianID, activeOnly, buildStatisticsQueryFilter(c), page, pageSize)
	if err != nil {
		h.Error(c, err)
		return
	}
	h.Success(c, stats)
}

// GetAssessmentEntryStatistics 查询单个测评入口统计。
// @Summary 查询单个测评入口统计
// @Description 按入口 ID 查询入口快照、窗口计数和最近事件时间；仅 qs:admin 可访问
// @Tags Statistics
// @Accept json
// @Produce json
// @Param Authorization header string true "Bearer 用户令牌"
// @Param id path uint64 true "入口ID"
// @Param preset query string false "时间窗口预设：today/7d/30d"
// @Param from query string false "自定义开始日期，格式 YYYY-MM-DD"
// @Param to query string false "自定义结束日期，格式 YYYY-MM-DD"
// @Success 200 {object} core.Response{data=statistics.AssessmentEntryStatistics}
// @Failure 429 {object} core.ErrResponse
// @Router /api/v1/statistics/entries/{id} [get]
func (h *StatisticsHandler) GetAssessmentEntryStatistics(c *gin.Context) {
	ctx := c.Request.Context()
	orgID, err := h.RequireProtectedOrgID(c)
	if err != nil {
		h.Error(c, err)
		return
	}
	entryID, err := strconv.ParseUint(c.Param("id"), 10, 64)
	if err != nil {
		h.Error(c, errors.WithCode(code.ErrInvalidArgument, "invalid entry id: %s", c.Param("id")))
		return
	}
	stats, err := h.readService.GetAssessmentEntryStatistics(ctx, orgID, entryID, buildStatisticsQueryFilter(c))
	if err != nil {
		h.Error(c, err)
		return
	}
	h.Success(c, stats)
}

// GetCurrentClinicianOverview 查询当前从业者统计概览。
// @Summary 查询当前从业者统计概览
// @Description 查询当前后台操作者绑定从业者的统计概览
// @Tags Statistics
// @Accept json
// @Produce json
// @Param Authorization header string true "Bearer 用户令牌"
// @Param preset query string false "时间窗口预设：today/7d/30d"
// @Param from query string false "自定义开始日期，格式 YYYY-MM-DD"
// @Param to query string false "自定义结束日期，格式 YYYY-MM-DD"
// @Success 200 {object} core.Response{data=statistics.ClinicianStatistics}
// @Failure 429 {object} core.ErrResponse
// @Router /api/v1/statistics/clinicians/me/overview [get]
func (h *StatisticsHandler) GetCurrentClinicianOverview(c *gin.Context) {
	ctx := c.Request.Context()
	orgID, operatorUserID, err := h.RequireProtectedScope(c)
	if err != nil {
		h.Error(c, err)
		return
	}
	stats, err := h.readService.GetCurrentClinicianStatistics(ctx, orgID, operatorUserID, buildStatisticsQueryFilter(c))
	if err != nil {
		h.Error(c, err)
		return
	}
	h.Success(c, stats)
}

// ListCurrentClinicianEntryStatistics 查询当前从业者入口统计列表。
// @Summary 查询当前从业者入口统计列表
// @Description 查询当前后台操作者绑定从业者的测评入口统计列表
// @Tags Statistics
// @Accept json
// @Produce json
// @Param Authorization header string true "Bearer 用户令牌"
// @Param preset query string false "时间窗口预设：today/7d/30d"
// @Param from query string false "自定义开始日期，格式 YYYY-MM-DD"
// @Param to query string false "自定义结束日期，格式 YYYY-MM-DD"
// @Param page query int false "页码"
// @Param page_size query int false "每页数量"
// @Success 200 {object} core.Response{data=statistics.AssessmentEntryStatisticsList}
// @Failure 429 {object} core.ErrResponse
// @Router /api/v1/statistics/clinicians/me/entries [get]
func (h *StatisticsHandler) ListCurrentClinicianEntryStatistics(c *gin.Context) {
	ctx := c.Request.Context()
	orgID, operatorUserID, err := h.RequireProtectedScope(c)
	if err != nil {
		h.Error(c, err)
		return
	}
	page, pageSize, err := h.parsePage(c)
	if err != nil {
		h.Error(c, err)
		return
	}
	stats, err := h.readService.ListCurrentClinicianEntryStatistics(ctx, orgID, operatorUserID, buildStatisticsQueryFilter(c), page, pageSize)
	if err != nil {
		h.Error(c, err)
		return
	}
	h.Success(c, stats)
}

// GetCurrentClinicianTesteeSummary 查询当前从业者受试者摘要。
// @Summary 查询当前从业者受试者摘要
// @Description 查询当前后台操作者绑定从业者可访问受试者摘要统计
// @Tags Statistics
// @Accept json
// @Produce json
// @Param Authorization header string true "Bearer 用户令牌"
// @Param preset query string false "时间窗口预设：today/7d/30d"
// @Param from query string false "自定义开始日期，格式 YYYY-MM-DD"
// @Param to query string false "自定义结束日期，格式 YYYY-MM-DD"
// @Success 200 {object} core.Response{data=statistics.ClinicianTesteeSummaryStatistics}
// @Failure 429 {object} core.ErrResponse
// @Router /api/v1/statistics/clinicians/me/testees-summary [get]
func (h *StatisticsHandler) GetCurrentClinicianTesteeSummary(c *gin.Context) {
	ctx := c.Request.Context()
	orgID, operatorUserID, err := h.RequireProtectedScope(c)
	if err != nil {
		h.Error(c, err)
		return
	}
	stats, err := h.readService.GetCurrentClinicianTesteeSummary(ctx, orgID, operatorUserID, buildStatisticsQueryFilter(c))
	if err != nil {
		h.Error(c, err)
		return
	}
	h.Success(c, stats)
}

// GetTesteePeriodicStatistics 查询受试者周期项目统计。
// @Summary 查询受试者周期项目统计
// @Description 查询受试者周期计划项目、周次任务和完成情况；后台访问范围按 ClinicianTesteeRelation 收口
// @Tags Statistics
// @Accept json
// @Produce json
// @Param Authorization header string true "Bearer 用户令牌"
// @Param testee_id path uint64 true "受试者ID"
// @Success 200 {object} core.Response{data=response.PeriodicStatsResponse}
// @Failure 429 {object} core.ErrResponse
// @Router /api/v1/statistics/testees/{testee_id}/periodic [get]
func (h *StatisticsHandler) GetTesteePeriodicStatistics(c *gin.Context) {
	ctx := c.Request.Context()
	testeeID, err := strconv.ParseUint(c.Param("testee_id"), 10, 64)
	if err != nil {
		h.Error(c, errors.WithCode(code.ErrInvalidArgument, "invalid testee id: %s", c.Param("testee_id")))
		return
	}
	orgID, operatorUserID, err := h.RequireProtectedScope(c)
	if err != nil {
		h.Error(c, err)
		return
	}
	if err := h.testeeAccessService.ValidateTesteeAccess(ctx, orgID, operatorUserID, testeeID); err != nil {
		h.Error(c, err)
		return
	}
	stats, err := h.periodicStatsService.GetPeriodicStats(ctx, orgID, testeeID)
	if err != nil {
		h.Error(c, err)
		return
	}
	h.Success(c, response.NewPeriodicStatsResponse(stats))
}

// BatchContentStatistics 批量查询统一内容统计。
// @Summary 批量查询统一内容统计
// @Description 按内容 type + code 批量查询测评形成数、完成数和完成率；questionnaire 需要 manage_questionnaires，scale 需要 manage_assessment_models，混合请求需要同时具备两项权限
// @Tags Statistics
// @Accept json
// @Produce json
// @Param Authorization header string true "Bearer 用户令牌"
// @Param request body contentBatchRequest true "内容标识列表"
// @Success 200 {object} core.Response{data=statistics.ContentBatchStatisticsResponse}
// @Failure 400 {object} core.ErrResponse
// @Failure 403 {object} core.ErrResponse
// @Failure 429 {object} core.ErrResponse
// @Router /api/v1/statistics/contents/batch [post]
func (h *StatisticsHandler) BatchContentStatistics(c *gin.Context) {
	var req contentBatchRequest
	if !h.bindJSON(c, &req) {
		return
	}
	orgID, err := h.RequireProtectedOrgID(c)
	if err != nil {
		h.Error(c, err)
		return
	}
	snapshot, _ := authzApp.FromContext(c.Request.Context())
	access := statisticsApp.ContentStatisticsAccess{
		Questionnaire: authzApp.SnapshotSatisfiesCapability(snapshot, authzApp.CapabilityManageQuestionnaires),
		Scale:         authzApp.SnapshotSatisfiesCapability(snapshot, authzApp.CapabilityManageAssessmentModels),
	}
	stats, err := h.readService.GetContentBatchStatistics(c.Request.Context(), orgID, req.Items, access)
	if err != nil {
		h.Error(c, err)
		return
	}
	h.Success(c, stats)
}

// ============= 定时任务 API =============

// SyncDailyStatistics 同步每日统计（内部系统动作）
// @Summary 同步每日统计
// @Description 从 MySQL 原始表重建每日统计；仅 qs:admin 可访问
// @Tags Statistics-Sync
// @Accept json
// @Produce json
// @Param Authorization header string true "Bearer 用户令牌（或内部调用token）"
// @Param start_date query string false "补算开始日期（格式：YYYY-MM-DD）"
// @Param end_date query string false "补算结束日期（格式：YYYY-MM-DD，包含当天）"
// @Success 200 {object} core.Response
// @Failure 429 {object} core.ErrResponse
// @Router /internal/v1/statistics/sync/daily [post]
func (h *StatisticsHandler) SyncDailyStatistics(c *gin.Context) {
	ctx := c.Request.Context()
	logger.L(ctx).Infow("同步每日统计", "action", "sync_daily_statistics")
	orgID, err := h.RequireProtectedOrgID(c)
	if err != nil {
		h.Error(c, err)
		return
	}

	opts, err := parseStatisticsSyncDateRange(c)
	if err != nil {
		h.Error(c, err)
		return
	}

	if err := h.syncService.SyncDailyStatistics(ctx, orgID, opts); err != nil {
		logger.L(ctx).Errorw("同步每日统计失败",
			"action", "sync_daily_statistics",
			"org_id", orgID,
			"error", err.Error(),
		)
		h.Error(c, err)
		return
	}
	h.governance().TriggerStatisticsWarmup(ctx, orgID, "sync_daily_statistics")

	h.Success(c, gin.H{"message": "每日统计同步完成"})
}

// SyncOrgSnapshotStatistics 同步机构总览快照（内部系统动作）
// @Summary 同步机构总览快照
// @Description 从 MySQL 原始表重建 statistics_org_snapshot；仅 qs:admin 可访问
// @Tags Statistics-Sync
// @Accept json
// @Produce json
// @Param Authorization header string true "Bearer 用户令牌（或内部调用token）"
// @Success 200 {object} core.Response
// @Failure 429 {object} core.ErrResponse
// @Router /internal/v1/statistics/sync/org-snapshot [post]
func (h *StatisticsHandler) SyncOrgSnapshotStatistics(c *gin.Context) {
	ctx := c.Request.Context()
	logger.L(ctx).Infow("同步机构总览快照", "action", "sync_org_snapshot_statistics")
	orgID, err := h.RequireProtectedOrgID(c)
	if err != nil {
		h.Error(c, err)
		return
	}

	if err := h.syncService.SyncOrgSnapshotStatistics(ctx, orgID); err != nil {
		logger.L(ctx).Errorw("同步机构总览快照失败",
			"action", "sync_org_snapshot_statistics",
			"org_id", orgID,
			"error", err.Error(),
		)
		h.Error(c, err)
		return
	}
	h.governance().TriggerStatisticsWarmup(ctx, orgID, "sync_org_snapshot_statistics")

	h.Success(c, gin.H{"message": "机构总览快照同步完成"})
}

// SyncPlanStatistics 同步计划统计（定时任务调用）
// @Summary 同步计划统计
// @Description 从 assessment_task 重建计划统计数据到 MySQL；仅 qs:admin 可访问
// @Tags Statistics-Sync
// @Accept json
// @Produce json
// @Param Authorization header string true "Bearer 用户令牌（或内部调用token）"
// @Success 200 {object} core.Response
// @Failure 429 {object} core.ErrResponse
// @Router /internal/v1/statistics/sync/plan [post]
func (h *StatisticsHandler) SyncPlanStatistics(c *gin.Context) {
	ctx := c.Request.Context()
	logger.L(ctx).Infow("同步计划统计", "action", "sync_plan_statistics")
	orgID, err := h.RequireProtectedOrgID(c)
	if err != nil {
		h.Error(c, err)
		return
	}

	if err := h.syncService.SyncPlanStatistics(ctx, orgID); err != nil {
		logger.L(ctx).Errorw("同步计划统计失败",
			"action", "sync_plan_statistics",
			"org_id", orgID,
			"error", err.Error(),
		)
		h.Error(c, err)
		return
	}
	h.governance().TriggerStatisticsWarmup(ctx, orgID, "sync_plan_statistics")

	h.Success(c, gin.H{"message": "计划统计同步完成"})
}

// RepairComplete marks an asynchronous cache repair as complete.
// @Summary 确认缓存修复完成
// @Description 内部治理回调，确认指定组织的缓存修复任务完成。
// @Tags Statistics-Sync
// @Accept json
// @Produce json
// @Param request body statistics.RepairCompleteRequest true "修复完成事件"
// @Success 200 {object} core.Response
// @Failure 400 {object} core.ErrResponse
// @Router /internal/v1/cache/governance/repair-complete [post]
func (h *StatisticsHandler) RepairComplete(c *gin.Context) {
	ctx := c.Request.Context()
	orgID, err := h.RequireProtectedOrgID(c)
	if err != nil {
		h.Error(c, err)
		return
	}

	var req statisticsApp.RepairCompleteRequest
	if !h.bindJSON(c, &req) {
		return
	}
	if err := h.governance().HandleRepairComplete(ctx, orgID, req); err != nil {
		h.Error(c, err)
		return
	}

	h.Success(c, gin.H{"message": "repair complete hook accepted"})
}

// WarmupTargets 手工触发缓存预热目标（内部治理动作）
// @Summary 手工触发缓存预热
// @Description operating 后台通过 BFF 代理调用，按 target 列表同步触发缓存预热并返回逐项结果；仅 qs:admin 可访问
// @Tags Statistics-Sync
// @Accept json
// @Produce json
// @Param Authorization header string true "Bearer 用户令牌（或内部调用token）"
// @Param request body cachetarget.ManualWarmupRequest true "预热目标列表"
// @Success 200 {object} core.Response{data=cachemodel.ManualWarmupResult}
// @Failure 400 {object} core.ErrResponse
// @Failure 429 {object} core.ErrResponse
// @Router /internal/v1/cache/governance/warmup-targets [post]
func (h *StatisticsHandler) WarmupTargets(c *gin.Context) {
	ctx := c.Request.Context()
	orgID, err := h.RequireProtectedOrgID(c)
	if err != nil {
		h.Error(c, err)
		return
	}
	var req cachetarget.ManualWarmupRequest
	if !h.bindJSON(c, &req) {
		return
	}

	var result *cachemodel.ManualWarmupResult
	result, err = h.governance().HandleManualWarmup(ctx, orgID, req)
	if err != nil {
		h.Error(c, err)
		return
	}
	h.Success(c, result)
}

// CacheGovernanceStatus returns the cache-governance runtime status.
// @Summary 查询缓存治理状态
// @Description 返回缓存治理任务及运行时状态，仅内部管理员可访问。
// @Tags Statistics-Sync
// @Produce json
// @Success 200 {object} core.Response
// @Router /internal/v1/cache/governance/status [get]
func (h *StatisticsHandler) CacheGovernanceStatus(c *gin.Context) {
	result, err := h.governance().GetStatus(c.Request.Context())
	if err != nil {
		h.Error(c, err)
		return
	}
	h.Success(c, result)
}

// CacheGovernanceHotset returns the active cache hotset.
// @Summary 查询缓存热集
// @Description 按 kind 和 limit 查询当前缓存热集，仅内部管理员可访问。
// @Tags Statistics-Sync
// @Produce json
// @Param kind query string false "缓存类型"
// @Param limit query int false "最大条数"
// @Success 200 {object} core.Response
// @Router /internal/v1/cache/governance/hotset [get]
func (h *StatisticsHandler) CacheGovernanceHotset(c *gin.Context) {
	result, err := h.governance().GetHotset(c.Request.Context(), c.Query("kind"), c.Query("limit"))
	if err != nil {
		h.Error(c, err)
		return
	}
	h.Success(c, result)
}

func (h *StatisticsHandler) governance() statisticsApp.GovernanceFacade {
	if h == nil {
		return statisticsApp.NewGovernanceFacade("apiserver", nil, nil)
	}
	if h.governanceFacade == nil {
		h.governanceFacade = statisticsApp.NewGovernanceFacade("apiserver", h.warmupCoordinator, h.cacheGovernanceStatusService)
	}
	return h.governanceFacade
}